refactor(todo): centralize semantic read policy
