+++ loopx/control_plane/agents/agent_lane_recommendation.py
@@ -11,7 +11,7 @@ from ..todos.contract import (
     normalize_todo_claimed_by,
     normalize_todo_id,
 )
-from ..todos.projection import todo_item_is_due_monitor
+from ..todos.todo_semantics import todo_item_is_due_monitor
 from ..todos.summary_item import compact_todo_summary_item
 from ..work_items.primary_action import protocol_action_text
 from ..work_items.work_lane import (
