+++ loopx/control_plane/agents/agent_scope.py
@@ -27,7 +27,7 @@ from ..todos.contract import (
 )
 from ..todos.handoff_gate import HandoffGateState
 from ..todos.resume_planning import project_todo_resume_planning
-from ..todos.projection import (
+from ..todos.todo_semantics import (
     todo_item_claimed_by_agent_or_unclaimed,
     todo_item_excludes_agent,
     todo_item_is_actionable_open,
