+++ loopx/extensions/lark/presentation/kanban.py
@@ -2196,7 +2196,7 @@ def sync_loopx_todos_to_lark_kanban(
     from ....capabilities.issue_fix.outcome_projection import (
         build_issue_fix_outcome_collection_from_domain_state,
     )
-    from ....control_plane.todos.projection import todo_priority_label
+    from ....control_plane.todos.todo_semantics import todo_priority_label
     from ....todos import resolve_todo_state_path, section_bounds, todo_blocks
 
     resolved_project, resolved_state_file = resolve_todo_state_path(
